# perf: optimize Groovy-based LAL DSL with static compilation