--- a/dist-material/release-docs/LICENSE
+++ b/dist-material/release-docs/LICENSE
@@ -455,7 +455,7 @@ Apache 2.0 licenses
 ========================================
 echarts	5.0.2:	https://github.com/apache/echarts	Apache-2.0
 Material Icons 3.0.1 https://github.com/google/material-design-icons Apache-2.0
-groovy             3.0.3 https://github.com/apache/groovy Apache-2.0
+groovy             3.0.8 https://github.com/apache/groovy Apache-2.0
 
 ========================================
 BSD licenses