--- a/oap-server/analyzer/log-analyzer/src/main/java/org/apache/skywalking/oap/log/analyzer/dsl/spec/parser/AbstractParserSpec.java
+++ b/oap-server/analyzer/log-analyzer/src/main/java/org/apache/skywalking/oap/log/analyzer/dsl/spec/parser/AbstractParserSpec.java
@@ -18,26 +18,32 @@
 
 package org.apache.skywalking.oap.log.analyzer.dsl.spec.parser;
 
-import lombok.Getter;
-import lombok.Setter;
 import lombok.experimental.Accessors;
 import org.apache.skywalking.oap.log.analyzer.dsl.spec.AbstractSpec;
 import org.apache.skywalking.oap.log.analyzer.provider.LogAnalyzerModuleConfig;
 import org.apache.skywalking.oap.server.library.module.ModuleManager;
 
-@Accessors(fluent = true)
+@Accessors
 public class AbstractParserSpec extends AbstractSpec {
     /**
      * Whether the filter chain should abort when parsing the logs failed.
      *
-     * Failing to parse the logs means either parsing throws exceptions or the logs not matching the desired patterns.
+     * Failing to parse the logs means either parsing throws exceptions or the logs not matching the
+     * desired patterns.
      */
-    @Getter
-    @Setter
     private boolean abortOnFailure = true;
 
     public AbstractParserSpec(final ModuleManager moduleManager,
                               final LogAnalyzerModuleConfig moduleConfig) {
         super(moduleManager, moduleConfig);
+    }
+
+    @SuppressWarnings("unused") // used in user LAL scripts
+    public void abortOnFailure(final boolean abortOnFailure) {
+        this.abortOnFailure = abortOnFailure;
+    }
+
+    public boolean abortOnFailure() {
+        return this.abortOnFailure;
     }
 }