--- a/oap-server/analyzer/log-analyzer/src/main/java/org/apache/skywalking/oap/log/analyzer/dsl/spec/parser/JsonParserSpec.java
+++ b/oap-server/analyzer/log-analyzer/src/main/java/org/apache/skywalking/oap/log/analyzer/dsl/spec/parser/JsonParserSpec.java
@@ -18,28 +18,23 @@
 
 package org.apache.skywalking.oap.log.analyzer.dsl.spec.parser;
 
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.skywalking.oap.log.analyzer.provider.LogAnalyzerModuleConfig;
 import org.apache.skywalking.oap.server.library.module.ModuleManager;
 
 public class JsonParserSpec extends AbstractParserSpec {
-    private final GsonBuilder gsonBuilder;
-
-    private final Gson gson;
+    private final ObjectMapper mapper;
 
     public JsonParserSpec(final ModuleManager moduleManager,
                           final LogAnalyzerModuleConfig moduleConfig) {
         super(moduleManager, moduleConfig);
 
-        gsonBuilder = new GsonBuilder();
-
-        // We just create a gson instance in advance for now (for the sake of performance),
+        // We just create a mapper instance in advance for now (for the sake of performance),
         // when we want to provide some extra options, we'll move this into method "create" then.
-        gson = gsonBuilder.create();
+        mapper = new ObjectMapper();
     }
 
-    public Gson create() {
-        return gson;
+    public ObjectMapper create() {
+        return mapper;
     }
 }