--- a/oap-server/analyzer/log-analyzer/src/test/resources/log-mal-rules/placeholder.yaml
+++ b/oap-server/analyzer/log-analyzer/src/test/resources/log-mal-rules/placeholder.yaml
@@ -13,22 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-rules:
-  - name: example
-    dsl: |
-      filter {
-        text {
-          abortOnFailure false // for test purpose, we want to persist all logs
-          regexp $/(?s)(?<timestamp>\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2}.\d{3}) \[TID:(?<tid>.+?)] \[(?<thread>.+?)] (?<level>\w{4,}) (?<logger>.{1,36}) (?<msg>.+)/$
-        }
-        extractor {
-          metrics {
-            timestamp log.timestamp as Long
-            labels level: parsed.level, service: log.service, instance: log.serviceInstance
-            name "log_count"
-            value 1
-          }
-        }
-        sink {
-        }
-      }
+# Refer to examples in config-examples/log-mal.yaml