--- a/tools/dependencies/known-oap-backend-dependencies.txt
+++ b/tools/dependencies/known-oap-backend-dependencies.txt
@@ -45,7 +45,7 @@ flatbuffers-java-1.12.0.jar
 freemarker-2.3.28.jar
 graphql-java-8.0.jar
 graphql-java-tools-5.2.3.jar
-groovy-3.0.3.jar
+groovy-3.0.8.jar
 grpc-api-1.32.1.jar
 grpc-context-1.32.1.jar
 grpc-core-1.32.1.jar