@wu-sheng 2021-07-08
Please fix CI. And serialization has more performance impact for groovy, right?