@wu-sheng 2021-07-09
> > Is this static compilation available for runtime script update or new script as input?
> 
> > I am thinking to share the same core for #7168 requirement
> 
> 
> 
> Yes. Whenever a new LAL DSL script is created, we can compile it to Java-style static byte codes once, after that its execution performance is close to Java.
> 

Then my next question would be, if we use this as debug tool? Whether could cause metaspace oom? Or we need to set up previous dymanic engine for debug tool.